src/main/webapp/app/entities/countries/update/countries-update.component.spec.ts
jest.mock('@angular/router');

import { ComponentFixture, TestBed } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { FormBuilder } from '@angular/forms';
import { ActivatedRoute } from '@angular/router';
import { of, Subject } from 'rxjs';

import { CountriesService } from '../service/countries.service';
import { ICountries, Countries } from '../countries.model';

import { CountriesUpdateComponent } from './countries-update.component';

describe('Component Tests', () => {
  describe('Countries Management Update Component', () => {
    let comp: CountriesUpdateComponent;
    let fixture: ComponentFixture<CountriesUpdateComponent>;
    let activatedRoute: ActivatedRoute;
    let countriesService: CountriesService;

    beforeEach(() => {
      TestBed.configureTestingModule({
        imports: [HttpClientTestingModule],
        declarations: [CountriesUpdateComponent],
        providers: [FormBuilder, ActivatedRoute],
      })
        .overrideTemplate(CountriesUpdateComponent, '')
        .compileComponents();

      fixture = TestBed.createComponent(CountriesUpdateComponent);
      activatedRoute = TestBed.inject(ActivatedRoute);
      countriesService = TestBed.inject(CountriesService);

      comp = fixture.componentInstance;
    });

    describe('ngOnInit', () => {
      it('Should update editForm', () => {
        const countries: ICountries = { id: 456 };

        activatedRoute.data = of({ countries });
        comp.ngOnInit();

        expect(comp.editForm.value).toEqual(expect.objectContaining(countries));
      });
    });

    describe('save', () => {
      it('Should call update service on save for existing entity', () => {
        // GIVEN
        const saveSubject = new Subject<HttpResponse<Countries>>();
        const countries = { id: 123 };
        jest.spyOn(countriesService, 'update').mockReturnValue(saveSubject);
        jest.spyOn(comp, 'previousState');
        activatedRoute.data = of({ countries });
        comp.ngOnInit();

        // WHEN
        comp.save();
        expect(comp.isSaving).toEqual(true);
        saveSubject.next(new HttpResponse({ body: countries }));
        saveSubject.complete();

        // THEN
        expect(comp.previousState).toHaveBeenCalled();
        expect(countriesService.update).toHaveBeenCalledWith(countries);
        expect(comp.isSaving).toEqual(false);
      });

      it('Should call create service on save for new entity', () => {
        // GIVEN
        const saveSubject = new Subject<HttpResponse<Countries>>();
        const countries = new Countries();
        jest.spyOn(countriesService, 'create').mockReturnValue(saveSubject);
        jest.spyOn(comp, 'previousState');
        activatedRoute.data = of({ countries });
        comp.ngOnInit();

        // WHEN
        comp.save();
        expect(comp.isSaving).toEqual(true);
        saveSubject.next(new HttpResponse({ body: countries }));
        saveSubject.complete();

        // THEN
        expect(countriesService.create).toHaveBeenCalledWith(countries);
        expect(comp.isSaving).toEqual(false);
        expect(comp.previousState).toHaveBeenCalled();
      });

      it('Should set isSaving to false on error', () => {
        // GIVEN
        const saveSubject = new Subject<HttpResponse<Countries>>();
        const countries = { id: 123 };
        jest.spyOn(countriesService, 'update').mockReturnValue(saveSubject);
        jest.spyOn(comp, 'previousState');
        activatedRoute.data = of({ countries });
        comp.ngOnInit();

        // WHEN
        comp.save();
        expect(comp.isSaving).toEqual(true);
        saveSubject.error('This is an error!');

        // THEN
        expect(countriesService.update).toHaveBeenCalledWith(countries);
        expect(comp.isSaving).toEqual(false);
        expect(comp.previousState).not.toHaveBeenCalled();
      });
    });
  });
});
